+++ dart_sample/lib/sample_routes.dart
@@ -94,6 +94,7 @@ class SampleRoutes {
   static const String baseline = '/baseline';
   static const String layoutBuilder = '/layout-builder';
   static const String keyboardListener = '/keyboard-listener';
+  static const String debugPainting = '/debug-painting';
 }
 
 class SampleRouteDefinition {